feat: initial changes to accomodate for bribesProcessor
