+++ brownie-config.yml
@@ -9,7 +9,7 @@ autofetch_sources: True
 # require OpenZepplin Contracts
 dependencies:
   - OpenZeppelin/openzeppelin-contracts@3.4.0
-  - GalloDaSballo/fair-selling@rc.0
+  - GalloDaSballo/fair-selling@rc.2
   - OpenZeppelin/openzeppelin-contracts@4.5.0/
 
 
@@ -19,7 +19,7 @@ compiler:
     version: null
     remappings:
       - "@openzeppelin=OpenZeppelin/openzeppelin-contracts@3.4.0"
-      - "@seller=GalloDaSballo/fair-selling@rc.0"
+      - "@seller=GalloDaSballo/fair-selling@rc.2"
       - "@oz=OpenZeppelin/openzeppelin-contracts@4.5.0/contracts/"
