+++ contracts/seller/OnChainPricer.sol
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: MIT
+pragma solidity 0.8.10;
+
+import {OnChainPricingMainnet} from "@seller/contracts/OnChainPricingMainnet.sol";
+
+contract OnChainPricer is OnChainPricingMainnet {
+  /// Using inheritance to import release from other brownie project
+  /// DO NOT CHANGE THIS
+  /// Chanding this file breaks trust in the code (untested)
+}
